Formulario autores y libros

+++ L02P02_2022MV650_2022MV652/Models/pedido_encabezado.cs
@@ -8,6 +8,6 @@ namespace L02P02_2022VF650_2022MV652.Models
         public int id { get; set; }
         public int id_cliente { get; set; }
         public int cantidad_libros { get; set; }
-        public decimal total { get; set; }
+        public double total { get; set; }
     }
 }
